fix cooldown and energy types

+++ clans/src/main/java/me/mykindos/betterpvp/clans/champions/skills/types/EnergySkill.java
@@ -2,6 +2,6 @@ package me.mykindos.betterpvp.clans.champions.skills.types;
 
 public interface EnergySkill {
 
-    int getEnergy(int level);
+    float getEnergy(int level);
 
 }
